updated crate_status.py for frame version print, ddeleted crate_status_frame.py
